requirements.txt
nemoguardrails
langchain
langchain_groq
fastapi
uvicorn
python-dotenv
